Moved Sequence and SearchNeighbor classes.
